Deriving.Aeson.Stripped

Documentation

A layer around the deriving-aeson package, with the ability to strip one or more fields from the JSON output, and recover them when decoding using some specified defaults.

Examples

Setup

All the examples on this page are based on the following setup code:

>>> :set -XDataKinds
>>> :set -XDeriveGeneric
>>> :set -XGeneralizedNewtypeDeriving
>>> :set -XOverloadedStrings
>>> :set -XTypeApplications

>>> import Data.Aeson
>>> import Deriving.Aeson.Stripped
>>> import qualified Data.Set as Set
>>> import Data.Text (Text)

>>> :{
 newtype WrappedInt
     = WrappedInt Int
     deriving (FromJSON, Show, ToJSON)
:}

>>> :{
 data RecordTest = RecordTest
     { testNumber   :: {-# UNPACK #-} !Int
     , testNewtype  :: !WrappedInt
     , testString   :: String
     , testIsString :: !Text
     , testList     :: ![Int]
     , testIsList   :: (Set.Set Int)
     , testMonoid   :: ![Int]
     }
     deriving (Generic, Show)
:}

>>> :{
 type StrippedRecordFields =
     '[ RField "testNumber"   7
      , RField "testNewtype"  (Coerce 42 Int)
      , RField "testString"   "string"
      , RField "testIsString" (FromString "text")
      , RField "testList"     '[ 10, 11, 12 ]
      , RField "testIsList"   (FromList '[ 13, 14, 13 ])
      , RField "testMonoid"   Mempty
      ]
:}

>>> :{
 data NonRecordTest
     = NonRecordTest (Either String Int) (Maybe Int) ![Int] (Bool, Char, Int)
     deriving (Generic, Show)
:}

>>> :{
 type StrippedNonRecordFields =
     '[ CField 0 ('Left "test")
      , CField 2 (Pure 7)
      , CField 1 'Nothing
      , CField 3 '( 'False, "z", 42 )
      ]
:}

Stripping fields in a record value: ..

>>> :{
 encode
     $ StrippedJSON @StrippedRecordFields @'[]
     $ RecordTest 1 (WrappedInt 2) "s" "t" [1..3] (Set.fromList [4..6]) [7..9]
:}
"[]"

.. and recovering them when decoding using the specified defaults:

>>> :{
 fmap unStrippedJSON
     $ decode @(StrippedJSON StrippedRecordFields '[] RecordTest)
     $ encode
     $ StrippedJSON @StrippedRecordFields @'[]
     $ RecordTest 1 (WrappedInt 2) "s" "t" [1..3] (Set.fromList [4..6]) [7..9]
:}
Just (RecordTest {testNumber = 7, testNewtype = WrappedInt 42, testString = "string", testIsString = "text", testList = [10,11,12], testIsList = fromList [13,14], testMonoid = []})

Stripping fields in a non-record value: ..

>>> :{
 encode
     $ StrippedJSON @StrippedNonRecordFields @'[]
     $ NonRecordTest (Right 1) (Just 2) [3..5] (True, 'a', 6)
:}
"[]"

.. and recovering them when decoding using the specified defaults:

>>> :{
 fmap unStrippedJSON
     $ decode @(StrippedJSON StrippedNonRecordFields '[] NonRecordTest)
     $ encode
     $ StrippedJSON @StrippedNonRecordFields @'[]
     $ NonRecordTest (Right 1) (Just 2) [3..5] (True, 'a', 6)
:}
Just (NonRecordTest (Left "test") Nothing [7] (False,'z',42))

Specifying encoding / decoding options:

The second parameter to StrippedJSON works exactly the same as the only parameter to CustomJSON (from the deriving-aeson package).

>>> :{
 encode
     $ StrippedJSON @'[] @'[ FieldLabelModifier CamelToSnake ]
     $ RecordTest 1 (WrappedInt 2) "s" "t" [1..3] (Set.fromList [4..6]) [7..9]
:}
"{\"test_number\":1,\"test_newtype\":2,\"test_string\":\"s\",\"test_is_string\":\"t\",\"test_list\":[1,2,3],\"test_is_list\":[4,5,6],\"test_monoid\":[7,8,9]}"

Core type

newtype StrippedJSON (fds :: [Type]) (opts :: [Type]) a

A newtype wrapper which provides FromJSON / ToJSON instances with a specific set of Options (see the Deriving.Aeson module), and the ability to strip one or more fields from the JSON output, recovered at decode-time using some default values.

Constructors

StrippedJSON

Fields

unStrippedJSON :: a

Stripping fields

data RField (name :: Symbol) (def :: k)

A field to be stripped from record values, identified by its name.

The def value is used to recover the field at decode-time (see RecoverableValue).

data CField (position :: Nat) (def :: k)

A field to be stripped from non-record single-constructor values, identified by its zero-based position in the data constructor.

The def value is used to recover the field at decode-time (see RecoverableValue).

Recovering fields

class RecoverableValue (x :: k) (a :: Type) where

A default field value which can be recovered at decode-time.

Methods

recoverValue :: Proxy x -> a

Recovers a default field value from the type-level.

data Coerce (x :: k) (a :: Type)

Recovers a value by coerce-ing it from another RecoverableValue.

>>> recoverValue (Proxy @(Coerce 7 Int)) :: WrappedInt
WrappedInt 7

data FromString (s :: Symbol)

Recovers a String-like value using fromString from IsString.

>>> recoverValue (Proxy @(FromString "text")) :: Text
"text"

data FromList (xs :: [k])

Recovers a list-like value using fromList from IsList.

>>> recoverValue (Proxy @(FromList '[1,2,3])) :: [Int]
[1,2,3]

data Mempty

Recovers a Monoid value using mempty.

>>> recoverValue (Proxy @(Pure 1)) :: Maybe Int
Just 1

data Pure (x :: k)

Recovers an Applicative value using pure.

>>> recoverValue (Proxy @(Pure 1)) :: Maybe Int
Just 1

Re-exports

module Deriving.Aeson